added figure for visualisation
