+++ events.py
@@ -1,13 +1,17 @@
 from collections import deque
+from math import sqrt
 
-SWING_HIGH_W = 150000  # threshold for angular acceleration for swing detect in CU(conditional unit)
-STAB_LOW_W = 10000  # low threshold for angular velocity for stab detect in CU
-STAB_HIGH_A = 10000000  # threshold for acceleration for stab detect in CU
-HIT_HIGH_A = 85000000  # threshols for acceleration for hit detect in CU
+SWING_HIGH_W = 1000000  # threshold for angular acceleration for swing detect in CU(conditional unit)
+SWING_LOW_W = 200000
+STAB_LOW_W = 1000000  # low threshold for angular velocity for stab detect in CU
+STAB_HIGH_A = 20000000  # threshold for acceleration for stab detect in CU
+HIT_HIGH_A = 200000000  # threshols for acceleration for hit detect in CU
+SWING_HIGH_A = 21000000
 SWING_TIME = 10  # number of measurements to detect swing
 HIT_TIME = 10  # number of measurements to detect hit using acceleration
 STAB_TIME = 5  # number of measurements to detect stab
 HIT_PAUSE = 50  # minimal time pause between different hits
+SWING_LOW_A = 17000000
 
 
 def update_acc_data(parameters: dict, actions: dict, a_curr: int, time: int):
@@ -15,11 +19,23 @@ def update_acc_data(parameters: dict, actions: dict, a_curr: int, time: int):
     Function updates acceleration parameters: start ans state of acceleration rising
     :param parameters: dict with parameters
     :param a_curr: current acceleration data
-    :param action: list of actions states
+    :param actions: list of actions states
+    :param time time count
     :return:
     """
     if a_curr >= HIT_HIGH_A:
         parameters['a_start'] = time
+    if a_curr >= SWING_HIGH_A and parameters['a_swing'] == 0:
+        parameters['a_swing_start'] = time
+        parameters['a_swing'] = 1
+    if a_curr >= STAB_HIGH_A and parameters['a_stab'] == 0:
+        parameters['a_stab'] = 1
+        parameters['a_stab_start'] = time
+    if a_curr < STAB_HIGH_A:
+        parameters['a_stab'] = 0
+        actions['stab'] = 0
+    if a_curr < SWING_LOW_A:
+        parameters['a_swing'] = 0
 
 
 def update_gyro_data(parameters: dict, actions: dict, w_curr: int, time: int):
@@ -28,26 +44,28 @@ def update_gyro_data(parameters: dict, actions: dict, w_curr: int, time: int):
     :param parameters: dict with parameters
     :param actions: dict with states
     :param w_curr: current angular velocity
-    :param stab_low_w - level of low angular velocity
+    :param time time counter
     :return:
     """
     if parameters['w_prev'] < w_curr and parameters['w_rising'] == 0:
         parameters['w_rising'] = 1
         parameters['w_start'] = time
-    if parameters['w_prev'] > w_curr:
+    """if parameters['w_prev'] > w_curr:
         parameters['w_rising'] = 0
-        actions['swing'] = 0
+        actions['swing'] = 0"""
     if w_curr < STAB_LOW_W and parameters['w_low'] == 0:
         parameters['w_low_start'] = time
         parameters['w_low'] = 1
     if w_curr > STAB_LOW_W:
         parameters['w_low'] = 0
         actions['stab'] = 0
+    if w_curr < SWING_LOW_W:
+        parameters['w_rising'] = 0
 
 
 def check_hit_with_accelerometer_and_change(acc_data: deque, time: int, parameters: dict, hit: int) -> bool:
     """
-    this function detects hits using high acceleration above ACC_HIT_LEVEL
+    this function detects hits using high acceleration above HIT_HIGH_A
     :param acc_data: 10 last accelerometer measurements
     :param time: current time counter
     :param parameters:  parameters of system
@@ -55,6 +73,11 @@ def check_hit_with_accelerometer_and_change(acc_data: deque, time: int, paramete
     :return:
     """
     if (time - parameters['a_start']) < HIT_TIME:
+        """if hit == 0 and (not parameters['hit_starts'] or time - parameters['hit_starts'][-1] > HIT_PAUSE):
+            print('HIT! at %i' % time)
+            if not time in parameters['hit_starts']:
+                parameters['hit_starts'].append(time)
+            return 1"""
         change = 0
         for i in range(min(HIT_TIME - 1, len(acc_data) - 1)):
             mul = sum([acc_data[0][j] * acc_data[i + 1][j] for j in range(3)])
@@ -63,12 +86,12 @@ def check_hit_with_accelerometer_and_change(acc_data: deque, time: int, paramete
             if change > 0 and hit == 0 and (
                         not parameters['hit_starts'] or time - parameters['hit_starts'][-1] > HIT_PAUSE):
                 print('HIT! at %i' % time)
-                if not time in parameters['hit_starts']:
+                if time not in parameters['hit_starts']:
                     parameters['hit_starts'].append(time)
-                return 1
+                return True
         if change == 0:
-            return 0
-    return 0
+            return False
+    return False
 
 
 def check_hit_with_change(acc_data: deque, time: int, parameters, hit) -> bool:
@@ -79,6 +102,7 @@ def check_hit_with_change(acc_data: deque, time: int, parameters, hit) -> bool:
     :param acc_data: data with accelerometer measures
     :param time: current time counter
     :param parameters: list of current parameters
+    :param hit is True if hit started
     :return: 1 if hit else 0
     """
     change = 0
@@ -87,27 +111,61 @@ def check_hit_with_change(acc_data: deque, time: int, parameters, hit) -> bool:
         if mul < 0:
             change += 1
         if change > 0 and hit == 0:
-            return 1
+            print('HIT! at %i' % time)
+            if time not in parameters['hit_starts']:
+                parameters['hit_starts'].append(time)
+            return True
     if change == 0:
-        return 0
+        return False
+
+
+def check_new_swing(gyro_data, time, parameters, swing) -> bool:
+    """
+    function detects swing. Swing is detected if angular velocity rises during last 10 measurements
+    and angular acceleration     for this time is more then SWING_HIGH_W threshold
+    :param gyro_data:  queue with last ten gyro measurements
+    :param time: current time counter
+    :param parameters: dict with parameteres
+    :return: 1 if swing else 0
+    """
+    if not swing:
+        div = sum(
+            [(gyro_data[SWING_TIME - 1][i] - gyro_data[0][i]) * (gyro_data[SWING_TIME - 1][i] - gyro_data[0][i]) for i
+             in
+             [1, 2]])
+        if (parameters['w_rising'] and (time - parameters['w_start']) > SWING_TIME and div > SWING_HIGH_W) or (
+            parameters['a_swing'] and (time - parameters['a_swing_start'] > SWING_TIME)):
+            print('SWING started at %i' % time)
+            if not time in parameters['swing_starts']:
+                parameters['swing_starts'].append(time)
+            return True
+        return False
+    if parameters['w_rising'] == 0:
+        print('SWING ended at %i' % time)
+        return False
+    return True
 
 
 def check_swing(gyro_data, time, parameters) -> bool:
     """
-    function detects swing. Swing is detected if angular velocity rises during last 10 measurements and angular acceleration
-    for this time is more then SWING_HIGH_W threshold
+    function detects swing. Swing is detected if angular velocity rises during last 10 measurements
+    and angular acceleration     for this time is more then SWING_HIGH_W threshold
     :param gyro_data:  queue with last ten gyro measurements
     :param time: current time counter
     :param parameters: dict with parameteres
     :return: 1 if swing else 0
     """
     if parameters['w_rising'] and (time - parameters['w_start']) > SWING_TIME:
-        div = sum(
-            [(gyro_data[SWING_TIME - 1][i] - gyro_data[0][i]) * (gyro_data[SWING_TIME - 1][i] - gyro_data[0][i]) for i
-             in range(3)])
+        # div = sum([(gyro_data[SWING_TIME-1][i] - gyro_data[0][i]) * (gyro_data[SWING_TIME-1][i] - gyro_data[0][i]) for i in [1, 2]])
+        div = sqrt(gyro_data[SWING_TIME - 1][1] * gyro_data[SWING_TIME - 1][1] + gyro_data[SWING_TIME - 1][2] *
+                   gyro_data[SWING_TIME - 1][2])
+        div -= sqrt(gyro_data[0][1] * gyro_data[0][1] + gyro_data[0][2] * gyro_data[0][2])
         if div / SWING_TIME > SWING_HIGH_W:
-            return 1
-    return 0
+            print('SWING started at %i' % parameters['w_start'])
+            if not parameters['w_start'] in parameters['swing_starts']:
+                parameters['swing_starts'].append(parameters['w_start'])
+            return True
+    return False
 
 
 def check_stab(acc_data, gyro_data, time, parameters) -> bool:
@@ -120,7 +178,10 @@ def check_stab(acc_data, gyro_data, time, parameters) -> bool:
     :param parameters: parameters of system
     :return: 1 is stab else 0
     """
-    div = sum([(acc_data[-1][i] - acc_data[0][i]) * (acc_data[-1][i] - acc_data[0][i]) for i in range(3)])
-    if div > STAB_HIGH_A and parameters['w_low'] and time - parameters['w_low_start'] > STAB_TIME:
-        return 1
-    return 0
+    if parameters['a_stab'] and (time - parameters['a_stab_start']) > STAB_TIME and parameters['w_low']:
+        if (time - parameters['w_low_start']) > STAB_TIME:
+            print('RUN!!! at %i' % parameters['a_start'])
+            if not parameters['a_start'] in parameters['stab_starts']:
+                parameters['stab_starts'].append(parameters['a_start'])
+        return True
+    return False
